chore: Add missing dependencies from
The `<MultipleSelector>` component has a few effects that have missing dependencies. We can observe this by installing the `eslint-plugin-react-hooks` eslint plugin.

This commit adds the plugin to the project (to help catch and future dependency issues) and fixes the warnings in the component. By adding the missing dependencies to the effects, we make sure the effects are triggered again in the case of said props changing.

I've made these changes to my actively used version of this component and haven't observed any side-effects.